[rules] Key object files based on the hash of their compilation flags.
This is an alternative to keyingthem on instance name, but it still
requires that compilation rules be moved into the instance rules.mk.
A small perl script for md5sum-ing a line of text is included, as perl
is more standard on these systems than the assurance that something
named exactly 'md5sum' will be in the path. Stupid.

The same hash runs 13 times, and it can't really be optimized away: filenames
have to have the hash in them, and the flags aren't truly known until the
moment of compilation. This is down from 33-42 times in initial
prototyping. During parallel building, a hack is employed to ensure that
the hashes don't get generated for compilations we'll never be doing
(parallel compilation is re-entrant and re-reads instance/rules.mk :().